% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aligned_histones}
\alias{aligned_histones}
\title{Aligned Human Histones}
\format{
A \code{list} of three \code{AAStringSetList} objects, each containing five elements
corresponding to the five histone families (H1, H2A, H2B, H3, H4):
\describe{
\item{unaligned}{\code{AAStringSetList} containing the unaligned histone sequences}
\item{msa}{\code{AAStringSetList} containing the aligned histone sequences after adding each sequence to the curated MSA profile of the corresponding family}
\item{msa_ref}{\code{AAStringSetList} containing the aligned reference sequences for each family}
}
}
\usage{
aligned_histones
}
\description{
A default call to \code{\link[=alignHistones]{alignHistones()}} retrieved all human histone variants from
UniProt (reviewed entries only, "2025-07-28 08:32:03 UTC") and added them to
the predefined (HistoneDB 2.0 (1)) MSA profile.
}
\examples{
# List all Histone H3 variants in the object
aligned_histones$unaligned$H3 |>
  names()
# Show the MSA of histone family H3
aligned_histones$msa$h3 |>
  print()
# Show the aligned reference sequences
aligned_histones$msa_ref |>
  purrr::walk(print)
}
\references{
(1) Draizen, E. J.; Shaytan, A. K.; Mariño-Ramírez, L.; Talbert, P. B.;
Landsman, D.; Panchenko, A. R. HistoneDB 2.0: A Histone Database with
Variants—an Integrated Resource to Explore Histones and Their Variants.
Database (Oxford) 2016, 2016, baw014. \url{https://doi.org/10.1093/database/baw014}.
}
\keyword{datasets}
