% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hptmMSA.R
\name{alignHistones}
\alias{alignHistones}
\title{Align Histone Sequences Using MAFFT}
\usage{
alignHistones(
  unaligned_histones = NULL,
  return_alignment = TRUE,
  output_path = "./out/msa/",
  overwrite = FALSE,
  use_profiles = TRUE,
  nondefault_refseq_names = NULL,
  return_only_original = TRUE,
  quiet = TRUE,
  num_cores = -1,
  ...
)
}
\arguments{
\item{unaligned_histones}{An optional \code{AAStringSetList} of histone sequences
to align, by family. If \code{NULL}, a default call to \code{\link[=histonesFromUniprot]{histonesFromUniprot()}}
is made, i.e., sequences from all five histone families (human and
reviewed) are retrieved.}

\item{return_alignment}{A \code{logical(1)} indicating whether to save the
alignment to \code{.fasta} files. Optional.}

\item{output_path}{A \code{character(1)} specifying the directory to save
alignment files. Optional.}

\item{overwrite}{A \code{logical(1)} indicating whether to overwrite existing
files at \code{output_path}. Optional.}

\item{use_profiles}{A \code{logical(1)} or \code{character()} indicating whether to
add sequences to existing histone MSA profiles. If \code{TRUE}, uses the curated
alignments from HistoneDB 2.0 (2). Can also be a \code{character()} specifying
paths to custom, pre-aligned \code{.fasta} files.}

\item{nondefault_refseq_names}{An optional \code{character()} of non-default
reference sequence names. If \code{NULL}, H1.1 and canonical H2A/H2B/H3/H4 are
used. Be aware that changing this can change hPTM location definition!}

\item{return_only_original}{A \code{logical(1)} indicating whether to return only
the sequences present in \code{unaligned_histones} in the MSA. Optional, only
used in profile mode.}

\item{quiet}{A \code{logical(1)} indicating whether to suppress MAFFT output.}

\item{num_cores}{An \code{integer(1)} specifying the number of CPU cores to use
during alignment. Optional, \code{-1} uses all cores.}

\item{...}{Additional arguments passed to MAFFT, overriding any defaults.}
}
\value{
A list containing the unaligned input sequences (\code{unaligned}), the
multiple sequence alignment (\code{msa}), and optionally the aligned reference
sequences (\code{msa_ref}) if \code{use_profiles} is not \code{FALSE}. Each list element
is a named \code{AAStringSetList}. An error is raised if MAFFT fails alignment
for any histone family.
}
\description{
Histone sequences can be aligned to prevent location mismatches between
corresponding hPTMs, e.g, H31S57 vs. H3CS56.
}
\details{
To ensure consistent and accurate alignment regardless of the underlying
sequences, this alignment is by default added into a predefined MSA profile
using the MAFFT \code{--add} functionality (1). This predefined MSA profile is by
default the curated alignment of the corresponding histone family retrieved
from HistoneDB 2.0 (2).

It is therefore highly recommended to leave these settings at default, so
that hPTM definition will be consistent, even when the sequences to align
change.
}
\references{
(1) Katoh, K.; Standley, D. M. MAFFT Multiple Sequence Alignment Software
Version 7: Improvements in Performance and Usability. Molecular Biology and
Evolution 2013, 30 (4), 772–780. \url{https://doi.org/10.1093/molbev/mst010}.
(2) Draizen, E. J.; Shaytan, A. K.; Mariño-Ramírez, L.; Talbert, P. B.;
Landsman, D.; Panchenko, A. R. HistoneDB 2.0: A Histone Database with
Variants—an Integrated Resource to Explore Histones and Their Variants.
Database (Oxford) 2016, 2016, baw014. \url{https://doi.org/10.1093/database/baw014}.
}
