% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{benchmark}
\alias{benchmark}
\title{hPTM Benchmark Data}
\format{
TODO A QFeatures with 7,240 rows and 60 columns:
\describe{
\item{assay}{contains the raw peptide intensities}
\item{rowData}{contains a variable "Proteins" with the protein accession and an variable ecoli to indicate if the protein is a spikin}
\item{colData}{contains a factor condition indicating the spike-in condition}
...
}
}
\source{
\url{https://www.ebi.ac.uk/pride/archive/projects/PXD009910}
}
\usage{
benchmark
}
\description{
A preprocessed \code{QFeatures} of hPTM data, functioning as a benchmark for
downstream analysis and statistics. The samples consist of commercial histone
extracts treated with a histone deacetylase (HDAC1) in a time-lapse design as
described in the reference below.
}
\examples{
benchmark |>
  show()
}
\references{
Clerck, L. D.; Willems, S.; Daled, S.; Puyvelde, B. V.;
Verhelst, S.; Corveleyn, L.; Deforce, D.; Dhaenens, M. An Experimental
Design to Extract More Information from MS-Based Histone Studies. Molecular
Omics 2021, 17 (6), 929–938. \url{https://doi.org/10.1039/D1MO00201E}.
}
\keyword{datasets}
