% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hptmMSA.R
\name{histonesFromUniprot}
\alias{histonesFromUniprot}
\title{Retrieve Histone Sequences From UniProt}
\usage{
histonesFromUniprot(
  histone_families = c("H1", "H2A", "H2B", "H3", "H4"),
  query = c("organism_id:9606", "reviewed:true"),
  name_field = "id",
  collapse = " AND "
)
}
\arguments{
\item{histone_families}{A \code{character()} of histone families to request. One
or more of \code{"H1"}, \code{"H2A"}, \code{"H2B"}, \code{"H3"}, or \code{"H4"}. Requests all
histone families by default.}

\item{query}{A \code{character()} containing all parts of the target query except
the "family" field. For query fields see https://www.uniprot.org/help/query-fields.
See also \code{\link[UniProt.ws:mapUniProt]{UniProt.ws::queryUniProt()}}.}

\item{name_field}{A \code{character(1)} specifying the UniProt field to use for
naming the sequences. For a full list of possible return fields see
https://www.uniprot.org/help/return_fields.}

\item{collapse}{A \code{character(1)} indicating how to collapse multiple query
clauses: either " OR " or " AND ". See also \code{\link[UniProt.ws:mapUniProt]{UniProt.ws::queryUniProt()}}.}
}
\value{
An \code{AAStringSetList} where each element corresponds to a histone
family and is an \code{AAStringSet} object of matching histone variant sequences.
}
\description{
Request histone families to the UniProt API and return an AAStringSetList of
all requested families.
}
