% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hptmUsage-package.R
\docType{package}
\name{hptmUsage-package}
\alias{hptmUsage}
\alias{hptmUsage-package}
\title{hptmUsage: Usage Analysis of Histone Post-Translational Modifications}
\description{
Preprocessing, visualization, and analysis (e.g., differential usage) of histone post-translational modifications (hPTMs). It builds on the 'msqrob2PTM' workflow to enable robust and statistically sound analysis of hPTM data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rualmey/hptmUsage}
  \item Report bugs at \url{https://github.com/rualmey/hptmUsage/issues}
}

}
\author{
\strong{Maintainer}: Ruben Almey \email{ruben.almey@ugent.be} (\href{https://orcid.org/0000-0001-6911-5904}{ORCID})

}
\keyword{internal}
