% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hptmUsage_data}
\alias{hptmUsage_data}
\title{Get Path to Data File(s)}
\usage{
hptmUsage_data(file = NULL)
}
\arguments{
\item{file}{Package data file name. If \code{NULL} (default), all data files are
listed.}
}
\description{
Convenience function for accessing data files (located in source at
\code{inst/extdata}) included with the hptmUsage package.
}
\examples{
hptmUsage_data()
hptmUsage_data("PXD009910_all_ID_pepion.csv")
}
