% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchHistones.R
\name{matchHistones}
\alias{matchHistones}
\alias{matchHistones,QFeatures,AAStringSetList-method}
\alias{matchHistones,SummarizedExperiment,AAStringSetList-method}
\title{Match Sequences to Histone Variants}
\usage{
matchHistones(object, matching_subject, ..., sequence_col = "sequence")

\S4method{matchHistones}{QFeatures,AAStringSetList}(
  object,
  matching_subject,
  i,
  progress = TRUE,
  drop_ambiguous = TRUE,
  sequence_col = "sequence"
)

\S4method{matchHistones}{SummarizedExperiment,AAStringSetList}(
  object,
  matching_subject,
  progress = TRUE,
  drop_ambiguous = TRUE,
  sequence_col = "sequence"
)
}
\arguments{
\item{object}{The \code{QFeatures} or \code{SummarizedExperiment} object in which the
sequences should be matched to histone variants.}

\item{matching_subject}{An \code{AAStringSetList} of histone families and
corresponding sequences. For example, the "unaligned" element resulting
from \code{\link[=alignHistones]{alignHistones()}}.}

\item{...}{Additional arguments passed to specific methods.}

\item{sequence_col}{The column name (\code{character(1)}) containing the sequences.}

\item{i}{The index (\code{integer(1)}) or name (\code{character(1)}) of the assay to be processed.}

\item{progress}{Show a progress bar? Defaults to \code{TRUE}.}

\item{drop_ambiguous}{Drop family- or position-ambiguous features? Defaults
to \code{TRUE}.}
}
\value{
A \code{QFeatures} or \code{SummarizedExperiment} (same as supplied) with
modified rowData of histone variant matches and the location.
}
\description{
Adds columns to the \code{rowData} of the given \code{QFeatures} or \code{SummarizedExperiment},
including checking for common errors.
}
\examples{
\dontrun{
ncbtoy |>
  matchHistones(aligned_histones$unaligned, 1)
}
\dontrun{
ncbtoy[[1]] |>
  matchHistones(aligned_histones$unaligned)
}
}
