% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgenesis.R
\name{readProgenesis}
\alias{readProgenesis}
\title{Read a Progenesis QIP Peptide Ion Data Export}
\usage{
readProgenesis(
  file,
  quant = "Raw abundance",
  generate_metadata = FALSE,
  overwrite_metadata = FALSE,
  simplify_column_names = TRUE
)
}
\arguments{
\item{file}{The path (\code{character(1)}) to the Progenesis QIP peptide ion data
\code{.csv} file. This file should contain all identified features, including
co-extracts, and at least the properties of feature number, charge,
protein, sequence, variable modifications, and a quantitative value (see
\code{quant}).}

\item{quant}{The quantitative value to use: \code{"Raw abundance"} (default),
\code{"Normalized abundance"}, or \code{"Intensity"}.}

\item{generate_metadata}{Generate a metadata \code{.csv} file that can be manually
edited (default \code{FALSE}, see also \code{overwrite_metadata})? Can also be a path
to the location where the metadata file will be written (e.g.,
\code{"./data/metadata.csv"}). Updated metadata can later be added to a
\code{QFeatures} object using \code{\link[=replace_metadata]{replace_metadata()}}.}

\item{overwrite_metadata}{Overwrite the metadata file if it already exists
(default \code{FALSE})?}

\item{simplify_column_names}{Remove pre-/suffixes from sample names (default
\code{TRUE})?}
}
\value{
A \code{QFeatures} containing the Progenesis QIP dataset.
}
\description{
A Progenesis QIP peptide ion data \code{.csv} file is cleaned up to make it
compatible with the \code{QFeatures} framework. This involves subsetting the
quantitative data to one type, e.g., raw abundances, sanitizing sample names,
and generating initial metadata from the Progenesis experimental design.
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
file.copy(hptmUsageData("all_ion_export.csv"), ".")
}
# Simplest case
readProgenesis("./all_ion_export.csv")

# Generate a metadata `.csv` file that can be manually edited
readProgenesis("./all_ion_export.csv", generate_metadata = TRUE)
# A path can also be supplied, instead of having the file generated next to
# the `.csv` source.
readProgenesis("./all_ion_export.csv", generate_metadata = "./metadata.csv")
# Note, if the metadata file already exists, nothing will be written unless
# `overwrite_metadata` is set to `TRUE`

# Use feature intensities instead of raw abundances
readProgenesis("./all_ion_export.csv", quant = "Intensity")

# Do not simplify (remove pre- and suffixes) sample names
readProgenesis("./all_ion_export.csv", simplify_column_names = FALSE)
\dontshow{
setwd(.old_wd)
}
}
\seealso{
The \code{\link[QFeatures:QFeatures-class]{QFeatures::QFeatures()}} class to read about how to manipulate
the resulting \code{QFeatures} object (if \code{return_qfeatures} was \code{TRUE}).
}
