% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceColData.R
\name{replaceColData}
\alias{replaceColData}
\alias{replaceColData,QFeatures_OR_SummarizedExperiment,character-method}
\alias{replaceColData,QFeatures_OR_SummarizedExperiment,tbl_df-method}
\title{Safely Replace \code{colData} of a QFeatures Object}
\usage{
replaceColData(object, colData, ...)

\S4method{replaceColData}{QFeatures_OR_SummarizedExperiment,character}(object, colData, custom_coltypes = NULL)

\S4method{replaceColData}{QFeatures_OR_SummarizedExperiment,tbl_df}(object, colData, custom_coltypes = NULL)
}
\arguments{
\item{object}{The QFeatures object in which to replace the colData.}

\item{colData}{TODO The path (\code{character(1)}) pointing to a colData \code{.csv} file. This file should contain at least the
columns of quantCols, original_name, group, include, and outlier.}

\item{coltypes}{Column types used by \code{readr::read_csv()}, also see \code{vignette("readr")} for more details.}
}
\value{
A \code{QFeatures} object with colData replaced.
}
\description{
TODO Description here. Included error checking...
}
\examples{
# TODO
NULL
}
