% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagContaminants.R
\name{tagContaminants}
\alias{tagContaminants}
\alias{tagContaminants,QFeatures-method}
\alias{tagContaminants,SummarizedExperiment-method}
\title{Tag Contaminant Proteins}
\usage{
tagContaminants(
  object,
  ...,
  library = c("universal", "cell_culture", "mouse_tissue", "rat_tissue",
    "neuron_culture", "stem_cell_culture"),
  sequence = "sequence"
)

\S4method{tagContaminants}{QFeatures}(
  object,
  i,
  library = c("universal", "cell_culture", "mouse_tissue", "rat_tissue",
    "neuron_culture", "stem_cell_culture"),
  sequence = "sequence"
)

\S4method{tagContaminants}{SummarizedExperiment}(
  object,
  library = c("universal", "cell_culture", "mouse_tissue", "rat_tissue",
    "neuron_culture", "stem_cell_culture"),
  sequence = "sequence"
)
}
\arguments{
\item{object}{The \code{QFeatures} or \code{SummarizedExperiment} object in which the
contaminants should be tagged.}

\item{...}{Additional arguments passed to specific methods.}

\item{library}{The name (\code{character(1)}) of the contaminant library to use.
Must be one of:
\itemize{
\item \code{"universal"} (default)
\item \code{"cell_culture"}
\item \code{"mouse_tissue"}
\item \code{"rat_tissue"}
\item \code{"neuron_culture"}
\item \code{"stem_cell_culture"}
}}

\item{sequence}{The name (\code{character(1)}) of the rowData column containing
(stripped) peptide sequences.}

\item{i}{The index (\code{integer()}) or name (\code{character()}) of the assay(s) to be processed.}
}
\value{
A \code{QFeatures} or \code{SummarizedExperiment} (same as supplied) with the rowData
containing one new \code{logical()} column "contaminant".
}
\description{
Use custom contaminant libraries from Frankenfield et al. (1), where histone
entries (Cont_A1A4R1, Cont_P62803, Cont_Q64475) were manually removed, to tag
contaminant proteins for later filtering. See
\url{https://github.com/HaoGroup-ProtContLib/Protein-Contaminant-Libraries-for-DDA-and-DIA-Proteomics}
for more information.
}
\examples{
# By default, contaminants will be tagged from the "universal" library
qf <- ncbtoy |>
  tagContaminants(i = "precursorRaw")
rowData(qf)
# but we can easily specify other libraries
qf_alt <- ncbtoy |>
  tagContaminants(i = "precursorRaw", library = "stem_cell_culture")
rowData(qf_alt)
# By default, contaminants will be tagged from the "universal" library
se <- ncbtoy[[1]] |>
  tagContaminants()
rowData(se)
# but we can easily specify other libraries
se_alt <- ncbtoy[[1]] |>
  tagContaminants(library = "stem_cell_culture")
rowData(se_alt)
}
\references{
(1) Frankenfield AM, Ni J, Ahmed M, Hao L. Protein contaminants matter:
building universal protein contaminant libraries for DDA and DIA proteomics.
Journal of proteome research. 2022 Jul 6;21(9):2104-13.
\url{https://pubs.acs.org/doi/full/10.1021/acs.jproteome.2c00145}.
}
