% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hptmMSA.R
\name{alignHistones}
\alias{alignHistones}
\title{Align histone sequences using MAFFT.}
\usage{
alignHistones(
  unaligned_histones = NULL,
  return_alignment = TRUE,
  output_path = "./out/msa/",
  overwrite = FALSE,
  use_profiles = TRUE,
  nondefault_refseq_names = NULL,
  return_only_original = TRUE,
  quiet = TRUE,
  num_cores = -1,
  ...
)
}
\arguments{
\item{unaligned_histones}{An \code{AAStringSetList} object containing sequences of
each histone family to align. If left empty, a default call to
\code{histonesFromUniprot()} is made, i.e., all five histone families from
human and reviewed.}

\item{return_alignment}{Whether or not to return the aligned fasta.}

\item{output_path}{A \code{character()} of the path where to store the aligned
fasta. Ignored if \code{return_alignment = FALSE}.}

\item{overwrite}{Overwrite aligned fasta if already present at
\code{output_path}? Ignored if \code{return_alignment = FALSE}.}

\item{use_profiles}{Use the curated alignments from HistoneDB 2.0 (1) as a
high-quality base MSA to which new sequences are added, ensuring consistent
and accurate alignment.

Can be \code{TRUE} to use profiles of H1, H2A, H2B, H3, H4, respectively, or can
be a \code{character()} specifying the histone families to use in order, e.g.,
\code{c("H3", "H4")} of specifying paths to custom pre-aligned \code{.fasta} files.

(1) Draizen, E. J.; Shaytan, A. K.; Mariño-Ramírez, L.; Talbert, P. B.;
Landsman, D.; Panchenko, A. R. HistoneDB 2.0: A Histone Database with
Variants—an Integrated Resource to Explore Histones and Their Variants.
Database (Oxford) 2016, 2016, baw014. https://doi.org/10.1093/database/baw014.}

\item{nondefault_refseq_names}{Names of the reference sequences for each histone family as
present in the profiles, with \code{NULL} using H1.1 and canonical H2A/H2B/H3/H4.
Be aware that changing this can change hPTM location definition! Ignored if
not using profiles.}

\item{return_only_original}{...}

\item{quiet}{...}

\item{num_cores}{...}

\item{...}{Additional arguments passed to MAFFT, overriding defaults.}
}
\value{
A named list of "unaligned" and "msa", each a named \code{AAStringSetList}
object, respectively.
}
\description{
Align histone sequences using MAFFT.
}
