% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate Differential Usage Analysis Report}
\usage{
generateReport(
  dataset,
  output_dir,
  overwrite = FALSE,
  histone_params = list(),
  msa_params = list(),
  mod_params = list(),
  contaminant_params = list(),
  usage_params = list(),
  ...
)
}
\arguments{
\item{dataset}{The hPTM dataset generated using \code{\link[=readProgenesis]{readProgenesis()}}.}

\item{output_dir}{The path where the final report should be saved (e.g., "./output").}

\item{overwrite}{If a report with the same name already exists at \code{output_dir},
should it be overwritten (\code{logical(1)}?}

\item{histone_params}{A \code{list()} of arguments passed to \code{\link[=histonesFromUniprot]{histonesFromUniprot()}}.}

\item{msa_params}{A \code{list()} of arguments passed to \code{\link[=alignHistones]{alignHistones()}} (except for "unaligned_histones").}

\item{mod_params}{A \code{list()} of arguments passed to \code{\link[=processMods]{processMods()}} (except for "object", "msa", or "i").}

\item{contaminant_params}{A \code{list()} of arguments passed to \code{\link[=tagContaminants]{tagContaminants()}} (except for "object" or "i").}

\item{usage_params}{A \code{list()} of arguments passed to \code{\link[=calculateUsage]{calculateUsage()}} (except for "object", "i", "name", or "target").}

\item{...}{Additional parameters passed to \code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}
}
\value{
Invisibly returns the final dataset, including aggregated hPTM assays, for example.
}
\description{
TODO
}
\details{
TODO
}
\examples{
\dontrun{
# TODO
# Usage level
...
}
}
