% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncbtoy}
\alias{ncbtoy}
\title{Example hPTM Data}
\format{
A QFeatures object with 1 assay named \code{precursorRaw}. This assay
contains 5663 features (peptide ions) and 10 samples from two conditions
('condition_A' and 'condition_B').
\describe{
\item{assay}{contains the raw peptide ion abundances}
\item{rowData}{contains 21 variables, including (most relevant) "protein", "sequence", and "mods"}
\item{colData}{contains 8 columns of sample metadata, including the "group" (biological factor), "ms_run"/"date_collected"/"prep_batch" (technical factors), and "include"/"outlier" (tags for analysis)}
}
}
\source{
\url{https://www.ebi.ac.uk/pride/archive/projects/PXD028162}
}
\usage{
ncbtoy
}
\description{
A minimally preprocessed \code{QFeatures} of hPTM data for testing and
demonstration purposes. The samples consist of bottom-up histone extracts
from naive (condition_A) and primed (condition_B) stem cell cultures as
described in the reference below.
}
\examples{
ncbtoy |>
  show()
}
\references{
Zijlmans, D. W.; Talon, I.; Verhelst, S.; Bendall, A.;
Van Nerum, K.; Javali, A.; Malcolm, A. A.; Van Knippenberg, S. S. F. A.;
Biggins, L.; To, S. K.; Janiszewski, A.; Admiraal, D.; Knops, R.;
Corthout, N.; Balaton, B. P.; Georgolopoulos, G.; Panda, A.; Bhanu, N. V.;
Collier, A. J.; Fabian, C.; Allsop, R. N.; Chappell, J.; Pham, T. X. A.;
Oberhuemer, M.; Ertekin, C.; Vanheer, L.; Athanasouli, P.; Lluis, F.;
Deforce, D.; Jansen, J. H.; Garcia, B. A.; Vermeulen, M.; Rivron, N.;
Dhaenens, M.; Marks, H.; Rugg-Gunn, P. J.; Pasque, V. Integrated
Multi-Omics Reveal Polycomb Repressive Complex 2 Restricts Human
Trophoblast Induction. Nat Cell Biol 2022, 24 (6), 858–871.
\url{https://doi.org/10.1038/s41556-022-00932-w}.
}
\keyword{datasets}
