% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProgenesis.R
\name{readProgenesis}
\alias{readProgenesis}
\title{Read a Progenesis QIP Peptide Ion Data Export}
\usage{
readProgenesis(
  file,
  quant = "Raw abundance",
  return_qfeatures = TRUE,
  simplify_column_names = TRUE,
  generate_metadata = TRUE,
  overwrite_metadata = FALSE
)
}
\arguments{
\item{file}{The path (\code{character(1)}) to the Progenesis QIP peptide ion data
\code{.csv} file. This file should contain all identified features, including
co-extracts, and at least the properties of feature number, charge,
protein, sequence, variable modifications, and a quantitative value (see
\code{quant}).}

\item{quant}{The quantitative value to use: \code{"Raw abundance"} (default),
\code{"Normalized abundance"}, or \code{"Intensity"}.}

\item{return_qfeatures}{Whether to return a \code{QFeatures} (\code{TRUE}, default)
object or a \code{tibble} (\code{FALSE}).}

\item{simplify_column_names}{Remove pre-/suffixes from sample names (default
\code{TRUE})?}

\item{generate_metadata}{Generate a metadata \code{.csv} file that can be manually
edited (default \code{TRUE}, see also \code{overwrite_metadata})? Can also be a path
to the location where the metadata file will be written. Updated metadata
can later be added to a \code{QFeatures} object using \code{\link[=replace_metadata]{replace_metadata()}}.}

\item{overwrite_metadata}{Overwrite the metadata file if it already exists
(default \code{FALSE})?}
}
\value{
A \code{QFeatures} or \code{tibble} (see \code{return_qfeatures}) containing the
Progenesis QIP dataset.
}
\description{
A Progenesis QIP peptide ion data \code{.csv} file is cleaned up to make it
compatible with the \code{QFeatures} framework. This involves subsetting the
quantitative data to one type, e.g., raw abundances, sanitizing sample names,
and generating initial metadata from the Progenesis experimental design.
}
\examples{
# Simplest case
}
\seealso{
The \code{\link[QFeatures:QFeatures-class]{QFeatures::QFeatures()}} class to read about how to manipulate
the resulting \code{QFeatures} object (if \code{return_qfeatures} was \code{TRUE}).
}
